Add expected license format to docs.
Also, strip out the https://... prefix for licenses, if it's there. A
lot of other tools that work with GCE use the URL form to reference
resources. Users may try to use the URL form of the license. This is a
simple thing we can do to improve cos-customizer's behavior when it's
given a license URL.

Also run mdformat on README.md.

Fixes #10.