/testing: Updated test timeouts (#23)
Currently, running all tests often results in some tests timing out. This is because a single Cloud Build project can only run at most 10 jobs in parallel, and time spent queued counts towards a job’s timeout. Let’s increase the timeout value to better handle the tests that we have. 7200s is 2 hours, which is a comfortably long time to run all of the tests.