tools/: added function ExtendOemPartition  (#25)
As the first step of "Verify Preloaded Data", we need to extend the OEM partition for storage of read-only user data. Given the situation that the OEM partition is surrounded by other partitions, the extension is divided into three steps.

1.moving the stateful partition towards the end of the disk (the stateful partition is the last partition on t
the disk)
2.moving the OEM partition to the original position of the stateful partition
3.extending the OEM partition and its file system
The interface of the function:
func ExtendOemPartition(disk string, statePartNum, oemPartNum int, oemSize string) error
oemSize can have unit like "100000B", "200K", "100M" or "2G". If no unit is provided, oemSize means the number of sectors (512B) for the extended OEM partition.

Testing:
A file in tools/partutil/disk_file is used as the simulation of a disk. When a test program starts, it will copy the file and work on it. Its size is 600K. It has three partitions as follows:
1.partition 8, OEM partition, 100K
2.partition 2, middle partition, 100K
3.partition 1, stateful partition, 100K
After running the test program for ExtendOemPartition, the copy of it will become:
1.empty space, 100K
2.partition 2, middle partition 100K
3.partition 8, OEM partition 200K
4.partition 1, stateful partition 100K
Since the functions work on partitions and file systems, tests need to be run with root permission ("sudo" is used in command executions).
The result of testing is verified by checking the file system space of the extended OEM partition and reading a file in each of the three partitions to make sure no data is corrupted.