Make sure that a unit exists before stopping it. (#27)
If we stop a unit that doesn't exist, `systemctl stop` will return a
non-zero exit code, resulting in a failed build. This causes issues when
we remove units in new milestones.

We use `systemctl is-active` to determine if we need to stop the unit or
not. This isn't the same thing as checking if the unit exists, but it's
good enough in this case because we really only want to stop the unit if
it is active.