Integrated extend-oem into cos-customizer (#26)
The OEM partition is 12MB in COS image, but we want it to be larger to store
read-only user data. So the extend_oem_partition program is integrated into
COS-Customizer.

Now users can add a -oem-size=${size} flage to the finish-image-build command
to set the OEM partition size. Since the OEM parition needs space for dm-verity hash tree,
its size is doubled. The disk size set with -disk-size-gb=${size} must be no smaller than 
--oem-size x 2 (rounded up to GB) + 10 to leave enough space for the OEM partition.

If the oem-size is set, startup.sh will create a unit that runs at shutdown time to execute a
binary program of extend_oem_partition.go and then reboot the instance. The binary file lives
in data/builtin_build_context/ .

Since we allow oem-size like "500M", and the OEM size is rounded up to GB, the extra space will
be taken by the stateful partition.