Integrate seal-oem into COS-Customizer
Bazel now will build a container image (about 33M)
with `veritysetup` stored in builtin_build_ctx.

A sub command `seal-oem` is added to configure
dm-verity of the OEM partition. When the subcommand
is run, a script name will be written to the state-file
that executes the script in builtin_build_ctx to run
the binary of seal-oem.

After sealing the OEM partition, the system will
not mount "/dev/sda8" after boot. Instead, "/dev/dm-1"
should be mounted at "/usr/share/oem/" manually.

If the subcommand is run, OEM parititon size will be
doubled to store the hash tree. If not, the OEM
partition size will be the same as user input.

Debian snapshot version is also updated to support
the dependencies of `veritysetup`.

The current method to save config.Build to config file
is to use `Save()`, but it doesn't clear the old file
before writing to it. So if the new JSON string of
config.Build is shorter, the tail of the old string will
still be in the config file, corrupting the JSON structure.
`func Update(configFile *os.File, buildConfig *Build) error`
is added to config/config.go. This method first clear the
config file and then use `Save()` to write to it.