added `disable-auto-update` subcommand
A subcommand `disable-auto-update` is added.
It will utilize tools/disable_systemd_service.go
to write a command to the grub to mask the systemd
service.

When this subcommand is run, it will write a script
name to the state file and then the VM will execute
the script in builtin_build_context to actually modify
the grub file.

When the `seal-oem` subcommand is run, it will
automatically disable the auto-update and disable the
usr-share-oem.mount to avoid the failure of mounting
/dev/sda8 because of the dm-verity on that partition.