reclaim space of sda3 when auto-update disabled
When the auto-update of COS is disabled, one of
the two root partitions will not be used. In order
to reclaim the space of the unused root partition,
`extend-oem.go` is changed into `handle-disk-layout.go`
that modifies the partitions on the disk.

Since the stateful partition is adjacent to sda3 (rootA),
we will change the priority of the partitions to make the
system boot from sda5 (rootB), and then shrink sda3 to
reclaim the space it has.

After shrinking sda3, we need to reboot the system
so that it can read the new partition table. And this needs
to happend before the `resize-disk` step to prevent the
stateful partition being extended. Therefore, `startup.sh`
is now divided into two parts, `prepare()` and `build()`.
`prepare()` runs before the `resize-disk` step in the Daisy
workflow and `build()` runs after that. And the Daisy
workflow is changed accordingly.

If the space of sda3 is reclaimed, 2046MB will be freed to
use by the OEM partition and the stateful partition. So the
limit of disk-size-gb is changed as well.

It takes about 54 minutes to run all test with `run_tests.sh`.