cloudbuild.yaml: release will not test tools/...
Currently the release workflow run `bazel test` on
all tests in the workspace. However, the tests
of programs related to disk partitions need to setup
loop devices. In cloud build VM, the loop device can
not be correctly set or it cannot be accessed directly
from an Ubuntu container. So we skip tests in tools/
to make sure the release workflow run successfully.

Tests in tools/ and tools/partutils/ should be
tested manually with root permission.