Add support for passing in an nvidia installer
For certain types of pre-release images, it is useful to be able to
choose the exact nvidia installer that we build with. This change allows
users to specify an nvidia installer in the 'install-gpu' step. The
installer must be specified in the '-version' flag, and if an installer
is specified, it must be available in the '-deps-dir' directory.