Add the "provisioner" command line tool
Currently, cos-customizer's on-instance provisioning behavior is
implemented in a startup script called "startup.sh". This script has
become pretty complicated in a number of ways, and is now very brittle.
Changes to that script need to be made very carefully and need to
account for subtle behaviors in bash and COS. Additionally, startup.sh
depends on COS-on-GCP-specific behaviors, like the get_metadata_value
script.

We can address these issues by moving almost all of the startup.sh
behaviors into a Go program. This will help with brittleness issues by:
- Having clear error handling behavior
- Having well-defined and explicit interfaces
- Removing dependencies on the base OS
- Removing the dependency on a Python container image
- Not needing to deal with bash `trap` builtin subtleties
- Having compile-time checks catch basic problems
- Giving us a standard way to write unit tests

The "provisioner" command line tool will be this Go program. The
interface will look like the following (subject to change as per code
reviews and implementation unknowns, of course):

./provisioner run --config=config.json
The run subcommand runs the provisioning behavior from a configuration
file. This configuration file is a replacement for startup.sh's state
file + GCE metadata flags.

./provisioner resume
The resume subcommand resumes an interrupted provisioning flow. This is
how we will handle provisioning steps that require rebooting the VM.

./provisioner status
The status subcommand gives the status of an in-progress provisioning
flow. This is how we know if there is an in-progress provisioning flow
or not, and helps the user know if `resume` makes sense.

All subcommands will take the --state-dir flag, which points to a
directory where the provisioner will store persistent state. This
replaces startup.sh's random files in /mnt/stateful_partition and
/var/lib/.cos-customizer.

This commit adds some bare files for the utility, and doesn't implement
any real behavior. Subsequent commits will implement the full behavior.