provisioner: Add InstallGPU step
The InstallGPU step is implemented using pretty much the same shell
script that we are already using. This is primarily because of the
requirement to install this script as part of installing GPU drivers.
It's simpler to have one script that is executed at both install time
and runtime than to have two nearly identical behaviors implemented in
different languages (Go and bash). I have considered making setup_gpu.sh
a Go program and installing that for runtime use, but implementing a
data dependency in that way seemed too complicated.