provisioner: Implement behavior for switching root device
COS Customizer currently offers repartitioning features that are
controlled through the disk size and OEM size options. Some of these
features depend on rebooting into COS's secondary update slot; in
particular, reclaiming disk space by disabling auto updates requires
that the system reboot into the secondary update slot. This change
implements behavior for rebooting into the secondary update slot. The
behavior is implemented in pretty much the same way it is implemented in
startup.sh. The biggest difference with startup.sh is how reboots are
handled.

In startup.sh, reboots are done in place as soon as the reboot is
identified as necessary. This introduced a few problems:
- There is a race between startup.sh and when startup.sh is killed
- trap statements sometimes execute
A few hacks were put in place in startup.sh to work around these
problems. This change addresses these problems by propagating reboot
requests to the user, through ErrRebootRequired. The user needs
to issue the reboot whenever they are ready for it.

In the specific context of COS Customizer, the user that needs to issue
the reboot will be a much thinner startup.sh. That script will focus
purely on dealing with run/resume/reboot semantics, and won't need to
worry about other provisioning details.