provisioner: Implement functionality for shrinking sda3
COS Customizer offers functionality for reclaiming /dev/sda3 space; it
does this by rebooting into sda5, minimizing the size of /dev/sda3, and
relocating other partitions to use that space. This change adds
functionality for minimizing the size of /dev/sda3.

The core of this behavior is already implemented in
partutil.MinimizePartition. Some small refactoring was needed to make
behavior that checks if a partition is already minimal available to the
provisioner.

./run_tests passes.