provisioner: Add WaitForDiskResize functionality
When repartitioning the disk, it is useful to enable arbitrary large
disk sizes. However, disk resize needs to occur at a specific time in
the repartitioning process to avoid partition relocation taking an
extremely long time. startup.sh currently handles this by expecting disk
resizing to occur during "wait_daisy_logging", which waits for a file to
exist in GCS. We handle it more intelligently here by waiting for the
disk to actually change in size.