start_image_build: Create the provisioner config
COS Customizer will utilize the provisioner tool by constructing a
provisioner config. To gradually and incrementally integrate the
provisioner, we will first introduce all of the logic to maintain
the provisioner config across all of the COS Customizer build steps,
and only after that remove other state that will be replaced by the
provisioner configuration (like the state file).

./run_tests passes.