run_script: Update provisioner config
The run_script step needs to update the provisioner config with the
provisioner's RunScript step.

As part of this change, I exported the RunScriptStep struct. This
simplifies the API between COS Customizer and the provisioner. In future
changes, I will make the structs for the rest of the provisioner steps
public as well.

./run_tests passes.