preloader: Add genrule for the "cidata" image
When we integrate the provisioner into COS Customizer, we will transfer
the provisioner to the preload VM through a special vfat file system
image with the CIDATA label. This serves two purposes:
- Keeping the provisioner on a separate disk prevents the provisioner
  binary from using disk resources on the preload VM, which could result
  in unintuitive ENOSPC errors in customer scripts. The provisioner
  binary bundles a few things, making it somewhat large (~50MB)
- The cloud-init "nocloud" datasource searches for configuration in a
  vfat file system with the CIDATA label. This disk will be used for
  delivering the cloud-init user-data to the VM for systems that do not
  support the GCE datasource (e.g. images for non-GCE platforms)

To build this "cidata" image, we need two new tools: mkfs.fat and mcopy.
We pull in the "dosfstools" and "mtools" third party deps to use these
tools. The Bazel build rules for these packages are put in
//src/third_party. This change adds a genrule that uses these tools to
build the "cidata" vfat image.