Add a metadata watcher binary
The current "daisy_ack" interface creates an awkward dependency
relationship between Daisy, the preload VM, the GCE metadata, and GCS.
This interface is solving a real problem, but that problem is specific
to Daisy. It doesn't make sense to implement this interface in the
provisioner, because the provisioner needs to also support non-Daisy
environments.

We can implement this interface in a simpler way by removing the
dependency on GCS and the host COS environment. In the startup script
that invokes the provisioner, we can first block on a metadata watcher
binary that waits for "daisy_ack" metadata to be set on the preload VM.
This metadata watcher binary can be delivered to the VM alongside the
provisioner.