Run unit tests with mtools installed
We use mtools to copy data onto the CIDATA disk image (which is a vfat
file system). Unit tests depend on mtools being installed in the
environment. Let's install mtools in the release workflow so that unit
tests pass there.

We install mtools by building a local container in the cloud build job
that is gcr.io/cloud-builders/bazel + mtools. We then use this container
in place of gcr.io/cloud-builders/bazel.