Break dependency on host's docker-credential-gcr
To allow running cos-customizer in environments that do not have
docker-credential-gcr installed, we can package docker-credential-gcr
with cos-customizer. We do this by embedding docker-credential-gcr into
the provisioner and writing it out to provisioner state when it is
needed.

Since the stateful partition on COS is not executable by default, we
mount a bind mount to execute docker-credential-gcr from.

I needed to re-order dependencies in WORKSPACE for the build to succeed
(we need a newer version of golang.org/x/sys for unix.MS_NOSYMFOLLOW).