Only remove customizer.service at shutdown time
Removing a systemd unit file changes the behavior of the unit. In
particular, the value of "TimeoutStartUSec" changes from "infinity"
(desired) to "1m 30s" (not desired) when the unit file is deleted.
Deleting customizer.service early in the preload process increases the
risk of TimeoutStartUSec kicking in and killing customizer.service. To
minimize this risk, let's only delete the unit file during system
shutdown (the stop behavior of customizer.service is invoked at shutdown
time).

A test is added to ensure that the unit file is indeed deleted. The test
can only be run on Ubuntu, because /etc is stateless on COS.